extract withPageContext
